analytics/views.py
from django.shortcuts import render
from rest_framework import generics
from .models import User, Business, Event
from .serializers import UserSerializer, BusinessSerializer, EventSerializer

class UserListView(generics.ListAPIView):
    queryset = User.objects.all()
    serializer_class = UserSerializer

class BusinessListView(generics.ListAPIView):
    queryset = Business.objects.all()
    serializer_class = BusinessSerializer

class EventListView(generics.ListAPIView):
    queryset = Event.objects.all()
    serializer_class = EventSerializer
